requirements.txt
beautifulsoup4==4.14.3
finance_datareader==0.9.42
numpy==1.23.5
pandas==1.5.3
PyMySQL==1.1.0
python-dotenv==1.2.1
Requests==2.32.5
SQLAlchemy==2.0.45
yfinance==0.2.66
